refactor: Sync and Async Code Re-use
Attempt to abstract sync and async requests by writing an abstraction
over the flow of data that encompasses a request. The requests fill out
a `RequestData` instance which then gets used as the source of future
transformations.

The login chain is a prime example of this. We define a series of steps
that are expressed as a series of sequences which will finally produce
the expected outcome.

The Session now produces a `Sequence` implementation which needs to be
driven by a client.

Unfortunately, due to lack of Async Trait support, the async
implementation is not as efficient as it could be. Attempt was made to
test the code on nightly, but ran into this bug
https://github.com/rust-lang/rust/pull/113108